+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Community Traits Ported (OpenMW)
 
+## 1.0.1
+
+- The Wretched and The Weird: Fixed Wretch's UI being dependant on Merlord's Backgrounds for no reason
+- The Wretched and The Weird: Fixed spell magicka calculations for Bog Witch and Blood Mage
+- All traits: Added Dagi-Raht to Khajiit whitelist
+- All traits: Added Malakh Orc to Orc whitelist
+
 ## 1.0
 
 Initial release